+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/service/command/feign/response/NaverUserInfo.java
@@ -1,0 +1,27 @@
+package com.pawith.authapplication.service.command.feign.response;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.Map;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class NaverUserInfo {
+    private static final String EMAIL = "email";
+    private static final String NICKNAME = "nickname";
+    private Map<String, String> response;
+
+    public NaverUserInfo(Map<String, String> response) {
+        this.response = response;
+    }
+
+    public String getNickname(){
+        return response.get(NICKNAME);
+    }
+
+    public String getEmail(){
+        return response.get(EMAIL);
+    }
+}
